requirements.txt
feedparser
markdown
python-dateutil